LayoutTests/accessibility/element-reflection-ariaerrormessage-expected.txt

Checks that element reflection is exposed to the a11y tree for 'ariaErrorMessageElements'

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS axTarget1.ariaErrorMessageElementAtIndex(0).isEqual(axError1) is true
PASS axTarget2.ariaErrorMessageElementAtIndex(0).isEqual(axError2) is true
PASS axTarget2.ariaErrorMessageElementAtIndex(0).isEqual(axError1) is true
PASS axInnerTarget.ariaErrorMessageElementAtIndex(0).isEqual(axError3) is true
PASS axTarget2.ariaErrorMessageElementAtIndex(0).isEqual(axError1) is true
PASS axTarget2.ariaErrorMessageElementAtIndex(1).isEqual(axError2) is true
PASS axTarget2.ariaErrorMessageElementAtIndex(2).isEqual(axError3) is true
PASS axTarget4.ariaErrorMessageElementAtIndex(0).isEqual(axErrorMessage4) is true
PASS axTarget5.ariaErrorMessageElementAtIndex(0).isEqual(axErrorMessage5) is true
PASS successfullyParsed is true

TEST COMPLETE
First error
Target 1
Second error
Target 2
Third error
Fourth error
Target 4
Fifth error
